+++ hata/discord/interaction/event_types/interaction_event.py
@@ -3,9 +3,9 @@ __all__ = ('InteractionEvent',)
 from scarletio import Future, export, future_or_timeout, include, shield
 
 from ...bases import DiscordEntity, EventBase
+from ...channel import create_partial_channel_from_id
 from ...core import (
-    APPLICATION_ID_TO_CLIENT, CHANNELS, GUILDS, INTERACTION_EVENT_MESSAGE_WAITERS, INTERACTION_EVENT_RESPONSE_WAITERS,
-    KOKORO
+    APPLICATION_ID_TO_CLIENT, GUILDS, INTERACTION_EVENT_MESSAGE_WAITERS, INTERACTION_EVENT_RESPONSE_WAITERS, KOKORO
 )
 from ...localizations.utils import get_locale
 from ...message import Message
@@ -538,7 +538,7 @@ class InteractionEvent(DiscordEntity, EventBase, immortal=True):
         -------
         channel : ``Channel``, `None`
         """
-        return CHANNELS.get(self.channel_id, None)
+        return create_partial_channel_from_id(self.channel_id, -1, self.guild_id)
     
     
     @property